@@ -1,35 +1,39 @@
 source 'https://rubygems.org'
+ruby "2.4.0"
 
-gem 'rails', '3.2.11'
+gem 'rails', '5.2.3'
 
-gem 'pg', group: 'production'
-gem 'mysql2', group: 'development'
+gem 'dotenv-rails', '~> 2.7', '>= 2.7.4'
+gem 'pg'#, group: 'production'
+#gem 'mysql2', group: 'development'
+gem 'bootsnap', '~> 1.4', '>= 1.4.4'
+gem 'listen', '~> 3.1', '>= 3.1.5'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
 
 # Gems used only for assets and not required
 # in production environments by default.
-group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+#group :assets do
+  #gem 'sass-rails',   '~> 3.2.3'
+  #gem 'coffee-rails', '~> 3.2.1'
 
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   # gem 'therubyracer'
 
-  gem 'uglifier', '>= 1.0.3'
-end
+gem 'uglifier', '>= 1.0.3'
+#end
 
-gem 'jquery-rails'
-gem "devise", "~> 2.0.4"
+#gem 'jquery-rails'
+gem "devise", "~> 4.6.2"
 gem "omniauth-twitter", "~> 0.0.8"
-gem "haml-rails", "~> 0.3.4"
-gem "twitter-bootstrap-rails", "~> 2.0.3"
+gem "haml-rails", "~> 2.0.1"
+gem "twitter-bootstrap-rails", "~> 4.0.0"
 gem "redcarpet", "~> 1.17.2"
-gem "nokogiri", "~> 1.5.2"
+gem "nokogiri", "~> 1.10.3"
 gem "pygmentize", "~> 0.0.3"
-gem "exception_notification", "~> 2.5.2"
+#gem "exception_notification", "~> 2.5.2"
 gem 'newrelic_rpm'
-gem "activeadmin", "~> 0.4.3"
-gem "meta_search",    '>= 1.1.0.pre'
-gem "therubyracer", "~> 0.11.2"
+gem "activeadmin", "~> 2.2.0"
+#gem "meta_search",    '>= 1.1.0.pre'
+#gem "therubyracer", "~> 0.11.2"
