@@ -1,4 +1,4 @@
-class AddTwitterFieldsToUser < ActiveRecord::Migration
+class AddTwitterFieldsToUser < ActiveRecord::Migration[5.2]
   def change
     add_column :users, :name, :string
     add_column :users, :twitter_oauth, :string
