@@ -1,4 +1,4 @@
-class CreateAdminNotes < ActiveRecord::Migration
+class CreateAdminNotes < ActiveRecord::Migration[5.2]
   def self.up
     create_table :admin_notes do |t|
       t.string :resource_id, :null => false
@@ -8,7 +8,6 @@ class CreateAdminNotes < ActiveRecord::Migration
       t.timestamps
     end
     add_index :admin_notes, [:resource_type, :resource_id]
-    add_index :admin_notes, [:admin_user_type, :admin_user_id]
   end
 
   def self.down
