@@ -1,5 +1,6 @@
 # Load the rails application
-require File.expand_path('../application', __FILE__)
+# require File.expand_path('../application', __FILE__)
+require_relative('application')
 
 # Initialize the rails application
-Myloveruby::Application.initialize!
+Rails.application.initialize!
